Drop support to laminas-servicemanager < 3.4
