+++ .gitignore
@@ -4,5 +4,6 @@
 /docs/html/
 /laminas-mkdoc-theme.tgz
 /laminas-mkdoc-theme/
+/.phpunit.result.cache
 /phpunit.xml
 /vendor/
