+++ composer.json
@@ -35,7 +35,7 @@
         "doctrine/annotations": "^1.10.4",
         "laminas/laminas-cache": "^2.9.0",
         "laminas/laminas-captcha": "^2.9.0",
-        "laminas/laminas-code": "3.4.99",
+        "laminas/laminas-code": "^3.5.0",
         "laminas/laminas-coding-standard": "^1.0.0",
         "laminas/laminas-escaper": "^2.6.1",
         "laminas/laminas-eventmanager": "^3.3.0",
@@ -50,23 +50,9 @@
         "phpspec/prophecy-phpunit": "^2.0",
         "phpunit/phpunit": "^9.4.2"
     },
-    "repositories": [
-        {
-            "type": "package",
-            "package": {
-                "name": "laminas/laminas-code",
-                "version": "3.4.99",
-                "dist": {
-                    "url": "https://github.com/Slamdunk/laminas-code/archive/php_80_support.zip",
-                    "type": "zip"
-                }
-            }
-        }
-
-    ],
     "suggest": {
         "laminas/laminas-captcha": "^2.9, required for using CAPTCHA form elements",
-        "laminas/laminas-code": "^3.4.1, required to use laminas-form annotations support",
+        "laminas/laminas-code": "^3.5, required to use laminas-form annotations support",
         "laminas/laminas-eventmanager": "^3.3, reuired for laminas-form annotations support",
         "laminas/laminas-i18n": "^2.10, required when using laminas-form view helpers",
         "laminas/laminas-recaptcha": "^3.2, in order to use the ReCaptcha form element",
@@ -76,10 +62,7 @@
     "autoload": {
         "psr-4": {
             "Laminas\\Form\\": "src/"
-        },
-        "files": [
-            "autoload/formElementManagerPolyfill.php"
-        ]
+        }
     },
     "autoload-dev": {
         "files": [
