+++ test/FormElementManagerFactoryTest.php
@@ -31,13 +31,7 @@ class FormElementManagerFactoryTest extends TestCase
         $elements = $factory($container, FormElementManager::class);
         $this->assertInstanceOf(FormElementManager::class, $elements);
 
-        if (method_exists($elements, 'configure')) {
-            // laminas-servicemanager v3
-            // $this->assertAttributeSame($container, 'creationContext', $elements);
-        } else {
-            // laminas-servicemanager v2
-            $this->assertSame($container, $elements->getServiceLocator());
-        }
+        // $this->assertAttributeSame($container, 'creationContext', $elements);
     }
 
     /**
@@ -54,27 +48,6 @@ class FormElementManagerFactoryTest extends TestCase
                 'test' => $element,
             ],
         ]);
-        $this->assertSame($element, $elements->get('test'));
-    }
-
-    /**
-     * @depends testFactoryReturnsPluginManager
-     */
-    public function testFactoryConfiguresPluginManagerUnderServiceManagerV2()
-    {
-        $container = $this->prophesize(ServiceLocatorInterface::class);
-        $container->willImplement(ContainerInterface::class);
-
-        $element = $this->prophesize(ElementInterface::class)->reveal();
-
-        $factory = new FormElementManagerFactory();
-        $factory->setCreationOptions([
-            'services' => [
-                'test' => $element,
-            ],
-        ]);
-
-        $elements = $factory->createService($container->reveal());
         $this->assertSame($element, $elements->get('test'));
     }
